dev-moonshot-api-lambda/requirements.txt
email-validator==1.1.3
uvicorn==0.17.0.post1
fastapi==0.92.0
python-dotenv==0.19.2
pyjwt==2.3.0
requests==2.27.1
mangum==0.12.4
pydantic==1.10.7
python-dateutil==2.8.2
pendulum==2.1.2
tomlkit==0.9.2
pytz==2022.1
openai==0.27.0
gradio_client==0.2.5
# google-cloud-aiplatform==1.25.0
# grpcio==1.54.0
# cygrpc
aiohttp==3.8.3
PyYAML==5.4.1
markdown==3.4.1
python-multipart==0.0.6